Access new column 'models' in DataFrame self
The current form `self.models = models` leads to a warning

UserWarning: Pandas doesn't allow columns to be created via a new attribute name - see https://pandas.pydata.org/pandas-docs/stable/indexing.html#attribute-access

because it is not the proper way to access a column that has be newly created. The correct way is `self['models'] = models`.