Passing parameters to parents of RunnableModel (#205)
Any additional parameters to the Model class are set as `params`. However, this is not the case for its child class RunnableModel, and therefore can't be used for child classes derived from it. To handle parameter input consistently they should be passed all the way up to the Model class. So, this PR adds `**params` to the RunnableModel init and to its super call.